requirements.txt
slackclient
certifi
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
json
